[tmux] unspecify TMUX default shell
The zsh version installed by default on macOS Monterey is 5.8.1
(/bin/zsh). This version will make the syntax highlighting of historical
information displayed by the arrow keys not properly colored. If you do
not specify a shell in TMUX, the system's login shell will be used,
which is good for sharing dotfiles on different platforms.